chore: move setQueryStoreFromAst to ovis.ts, add tests, and remove addStratifierToQuery
