+++ package.json
@@ -45,6 +45,7 @@
         "eslint-plugin-svelte": "^3.1.0",
         "globals": "^16.0.0",
         "husky": "^9.1.7",
+        "jsdom": "^26.1.0",
         "lint-staged": "^16.1.2",
         "prettier": "^3.5.3",
         "prettier-plugin-svelte": "^3.3.3",
